preserve size_t

+++ openexr-c/include/imf_deepframebuffer.h
@@ -3,6 +3,7 @@
 
 #include <imf_pixeltype.h>
 #include <stdbool.h>
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -46,9 +47,9 @@ OPENEXR_CPPMM_API unsigned int Imf_3_0__DeepSlice_ctor(
     Imf_DeepSlice_t * this_
     , Imf_PixelType type
     , char * base
-    , unsigned long xStride
-    , unsigned long yStride
-    , unsigned long sampleStride
+    , size_t xStride
+    , size_t yStride
+    , size_t sampleStride
     , int xSampling
     , int ySampling
     , double fillValue
